requirements.txt
aiohttp @ file:///C:/ci_311/aiohttp_1676432932774/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
async-timeout @ file:///C:/ci_311/async-timeout_1676431518331/work
attrs @ file:///C:/ci_311/attrs_1676422272484/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///C:/b/abs_13590mi9q9/croot/cryptography_1689373706078/work
frozenlist @ file:///C:/ci_311/frozenlist_1676428131576/work
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna @ file:///C:/ci_311/idna_1676424932545/work
multidict @ file:///C:/ci_311/multidict_1676428396308/work
oauthlib==3.2.2
plyer==2.1.0
protobuf==4.24.2
pyyaml==6.0
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.1.1
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dotenv==0.21.0
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
requests-oauthlib==1.3.1
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
yaml==0.2.5
yarl @ file:///C:/ci_311/yarl_1676432870023/work
